Implement Price repository with JPA integration
- Updated Price domain model to use Long for priceList to match database schema
- Enhanced PriceRepositoryPort interface with more specific methods:
  - findApplicablePriceAt: returns Optional<Price> with highest priority
  - findApplicablePricesAt: returns List<Price> for all applicable prices
- Implemented PriceRepositoryAdapter with proper JPA integration
- Added JPQL queries to JpaPriceRepository for finding prices at specific datetime
- Added toDomainModel() mapper to JpaPriceEntity for clean domain conversion

These changes complete the repository layer implementation, allowing the
application to properly retrieve price information based on product, brand,
and application datetime with priority handling.

+++ src/main/java/com/example/test/price/domain/models/Price.java
@@ -7,13 +7,13 @@ public class Price{
     Long brandId;
     LocalDateTime startDate;
     LocalDateTime endDate;
-    Integer priceList;
+    Long priceList;
     Long productId;
     Integer priority;
     BigDecimal price;
     String currency;
 
-    public Price(Long brandId, LocalDateTime startDate, LocalDateTime endDate, Integer priceList, Long productId, Integer priority, BigDecimal price, String currency) {
+    public Price(Long brandId, LocalDateTime startDate, LocalDateTime endDate, Long priceList, Long productId, Integer priority, BigDecimal price, String currency) {
         this.brandId = brandId;
         this.startDate = startDate;
         this.endDate = endDate;
@@ -36,7 +36,7 @@ public class Price{
         return endDate;
     }
 
-    public Integer getPriceList() {
+    public Long getPriceList() {
         return priceList;
     }
